requirements.txt
asgiref==3.9.1
attrs==25.3.0
Django==5.2.4
django-phonenumber-field==8.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
drf-yasg==1.21.10
inflection==0.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
packaging==25.0
phonenumbers==9.0.10
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.26.0
sqlparse==0.5.3
style==1.1.0
typing_extensions==4.14.1
update==0.0.1
uritemplate==4.2.0
